==> 05_teaserify.Rmd <==
---
title: "05_teaserify"
author: "Duc-Quang Nguyen"
date: "24 July 2016"
output: html_document
---

```{r yo, include=FALSE}
library(magick)

pic.path <- "teaser"
# find the available front image for the teaser -> define languages
languages <- gsub(".*_(.*)\\.png", "\\1", list.files(
  pic.path, pattern = paste0("^01_teaser_flickr_")))

pic.files <- list.files(pic.path, full.names = T, pattern = "^[1-6]")

last.image <- list.files(pic.path, full.names = T, pattern = "^logo")

sapply(languages, function(lang) {
  first.image <- list.files(pic.path, pattern = paste0("01_teaser_flickr_", lang, ".png"), full.names = T)
  
  pics <- image_scale(image_read(c(first.image, pic.files, last.image)) ,"1024x512") 
  animation1 <- image_animate(c(pics), fps = 0.4)
  image_write(animation1, paste0("teaser_flickrstory_", lang, ".gif"))
})


```
